implement CustomUserDetailsService class for handling loading user details from db & to validate JWTs and authenticate users.
